requirements.txt
django
gunicorn
redis
django-redis
celery
PyMySQL
mysqlclient
djangorestframework
django-filter
requests
ipython